Enable multiple roles per user. Closes #19
This allows, say, a National Finance Head to act as a Volunteer and
register donations in his own name. The server now returns an array of
roles which is handled by the updated client, instead of a single role,
to make this possible.

However, for a higher role user to act as a lower role user requires
that all intermmediary roles be filled as well. i.e., for a NFH to be
a volunteer, he will have to be a Fellow and a Coach as well to
preserve the chain of responsibility.